Mv make_dir() to helper.py & fix

+++ src/util/helper.py
@@ -95,8 +95,9 @@ class Helper:
                 file_name = get_file_name(r)
 
             file_path = os.path.join(folder, file_name)
-
+            print(os.path.exists(file_path))
             with open(file_path, 'wb') as f:
+                print('@@@@@@@@@@@@', file_path, r.status)
                 shutil.copyfileobj(r, f)
         finally:
             r.close()
@@ -362,13 +363,15 @@ class Helper:
                }       
 
     @staticmethod
-    def make_dir(path):
+    def make_dir(path, remake_if_exists=False):
         try:
             os.makedirs(path)
         except OSError:
-            print(path, ' already exists, remaking...')
-            shutil.rmtree(path)
-            os.makedirs(path)
+            print(path, ' already exists!')
+            if remake_if_exists:
+                print('Remaking...')
+                shutil.rmtree(path)
+                os.makedirs(path)
 
     @staticmethod
     def deintonate_txt(txt):
